Split standard suite test into three separate tests
