+++ tests/standard_suite_traces_cpu.cc
@@ -1,36 +1,6 @@
 #include <iostream>
 #include <vector>
 
-int cpu_stream_test() {
-  char *command;
-
-  int ret = asprintf(&command,
-      "../spatter -f "
-      "../../standard-suite/basic-tests/cpu-stream.json");
-  if (ret == -1 || system(command) != EXIT_SUCCESS) {
-    std::cerr << "Test failure on " << command << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  free(command);
-  return EXIT_SUCCESS;
-}
-
-int cpu_ustride_test() {
-  char *command;
-
-  int ret = asprintf(&command,
-      "../spatter -f "
-      "../../standard-suite/basic-tests/cpu-ustride.json");
-  if (ret == -1 || system(command) != EXIT_SUCCESS) {
-    std::cerr << "Test failure on " << command << std::endl;
-    return EXIT_FAILURE;
-  }
-
-  free(command);
-  return EXIT_SUCCESS;
-}
-
 int cpu_amg_test() {
   char *command;
 
@@ -90,12 +60,6 @@ int cpu_pennant_test() {
 int main(int argc, char **argv) {
   (void)argc;
   (void)argv;
-
-  if (cpu_stream_test() != EXIT_SUCCESS)
-    return EXIT_FAILURE;
-
-  if (cpu_ustride_test() != EXIT_SUCCESS)
-    return EXIT_FAILURE;
 
   if (cpu_amg_test() != EXIT_SUCCESS)
     return EXIT_FAILURE;
